% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{affect_lines_type}
\alias{affect_lines_type}
\title{affect line type}
\usage{
affect_lines_type(lt_lines, nb_type)
}
\arguments{
\item{lt_lines}{(list to named row): the line landscape whose affect types}

\item{nb_type}{(int) : number types are affect}
}
\value{
lt_lines (list to named row): the line landscape with types
}
\description{
Give a type at each line (ramdom)
}
