% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{affect_polygons_type}
\alias{affect_polygons_type}
\title{affect polygon type}
\usage{
affect_polygons_type(landscape, nb_type)
}
\arguments{
\item{landscape}{(SpatialPolygonsDataFrame): the landscape whose affect types}

\item{nb_type}{(int) : number types are affect}
}
\value{
landscape (SpatialPolygoneDataFrame): the landscape with types
}
\description{
Give a type at each polygon (ramdom)
}
