% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{all_effect}
\alias{all_effect}
\title{all effect}
\usage{
all_effect(nv_coords_point, ui_width, ui_heigth, d_sigma, nm_dist_element,
  nm_info_type, time_step)
}
\arguments{
\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{d_sigma}{(double) repulsif effect adaptor}

\item{nm_dist_element}{(NumericMatrix) all coordinates of segment 
Matrix N*5 (x1,y1,x2,y2,Id_type)}

\item{nm_info_type}{(NumericMatrix) descript each type
Matrix M*5 (Id, rep (-1) or attr(1), 
            alpha, beta, power)}

\item{time_step}{(double) step of time}

\item{ui_land_width}{(unsigned int) width of landscape.}

\item{ui_land_heigth}{(unsigned int) heigth of landscape.}
}
\value{
(NumericVector) effect in x and y (['x'];['y'])
}
\description{
Give complete effect of landscape on x and y
}
