% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{all_value}
\alias{all_value}
\title{all value}
\usage{
all_value(nv_coords_point, nm_coords_attrac, nm_coords_rep)
}
\arguments{
\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{nm_coords_attrac}{(NumericMatrix) all coordinates of attractive 
segment matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
alpha_a (double) maximum potential 
amplitude of attractivity
beta (double) spatial display of 
potential, and power (double)}

\item{nm_coords_rep}{(NumericMatrix) all coordinates of repulsif segment 
Matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
   alpha_a (double) maximum potential 
   amplitude of repulsivity
   beta (double) spatial display of 
   potential, and power (double)}
}
\value{
(double) potential value of landscape on x and y
}
\description{
Give complete potential value of landscape on x and y
}
