% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alpha_func}
\alias{alpha_func}
\title{alpha function}
\usage{
alpha_func(alpha1, alpha2, alpha3, t)
}
\arguments{
\item{alpha1}{(double) maximum potential amplitude}

\item{alpha2}{(double) spatial display of potential in time}

\item{alpha3}{(double) time when potential amplitude is maximum}
}
\value{
(double) miximum of potential at t
}
\description{
evolution of maximum of potential in time
}
