% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attractive_effect}
\alias{attractive_effect}
\title{attractive effect}
\usage{
attractive_effect(nm_coords_attrac, nv_coords_point)
}
\arguments{
\item{nm_coords_attrac}{(NumericMatrix) all coordinates of attractive 
segment matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
alpha_a (double) maximum potential 
amplitude of attractivity
beta (double) spatial display of 
potential, and power (double)}

\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}
}
\value{
(NumericVector) attractive effect in x and y (['x'];['y'])
}
\description{
Give effect on x and y of all attractive elements
}
