% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{border_effect}
\alias{border_effect}
\title{Border effect}
\usage{
border_effect(ui_width, ui_heigth, d_sigma, d_pos_x, d_pos_y)
}
\arguments{
\item{d_sigma}{(double) repulsif effect adaptor}

\item{d_pos_x}{(double) position in x}

\item{d_pos_y}{(double) position in y}

\item{ui_land_width}{(unsigned int) width of landscape.}

\item{ui_land_heigth}{(unsigned int) heigth of landscape.}

\item{ad_effect_xy}{(double[2]) effect in x and y}
}
\description{
This function calculate the effect of border on mouvement
}
