% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bound}
\alias{bound}
\title{bound}
\usage{
bound(val, bound_1, bound_2)
}
\arguments{
\item{val}{(double) value must be staked}

\item{bound_1}{(double) first bound}

\item{bound_2}{(double) second bound}
}
\value{
(double) staked value
}
\description{
stake value out
}
