% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{commun_coords}
\alias{commun_coords}
\title{commun coords}
\usage{
commun_coords(landscape, ids)
}
\arguments{
\item{landscape}{(SpatialPolygonsDataFrame): SIG}

\item{ids}{(tupple of int): id of two polygons}
}
\value{
(matrix X*2): coordonate of commun points
}
\description{
give commun coordonate between two polygons
}
