% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diffusion}
\alias{diffusion}
\title{diffusion}
\usage{
diffusion(d_sigma, time_step)
}
\arguments{
\item{d_sigma}{(double)}

\item{time_step}{(double) step of time}
}
\value{
(double) diffusion
}
\description{
calculate the diffusion
}
