% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{dist_expr_two_points}
\alias{dist_expr_two_points}
\title{Distance expression between two points}
\usage{
dist_expr_two_points(x0, y0, x1, y1)
}
\arguments{
\item{x0}{(float or characters): abscisse of first point of segment}

\item{y0}{(float or characters): ordonate of first point of segment}

\item{x1}{(float or characters): abscisse of second point of segment}

\item{y1}{(float or characters): ordonate of second point of segment}
}
\value{
string
}
\description{
Give expresssion of calcul of distance between two points
}
