% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance2point}
\alias{distance2point}
\title{Distance to point}
\usage{
distance2point(x, y, x1, y1)
}
\arguments{
\item{x}{(double) x of first point}

\item{y}{(double) y of first point}

\item{x1}{(vector) x1 of second point}

\item{y1}{(vector) y1 of second point}
}
\value{
(vector) distance to segment (dist),
                 derivate in x (dx),
                 derivate in y (dy)
}
\description{
Give distance a point to point and derivate in x and y
}
