% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{distance2poly}
\alias{distance2poly}
\title{distance point to polygon}
\usage{
distance2poly(x, y, land_poly, id)
}
\arguments{
\item{x}{(double) position in x}

\item{y}{(double) position in y}

\item{land_poly}{(SpatialPolygon): polygon landscape}

\item{Id}{: Id of polygon}
}
\value{
(vector) distance to segment (dist),
                 derivate in x (dx),
                 derivate in y (dy)
}
\description{
Give distance a point to polygon and derivate in x and y
}
