% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance2segment}
\alias{distance2segment}
\title{Distance to Segment}
\usage{
distance2segment(x, y, x1, y1, x2, y2)
}
\arguments{
\item{x}{(double) x of point}

\item{y}{(double) y of point}

\item{x1}{(vector) x1 of segment}

\item{y1}{(vector) y1 of segment}

\item{x2}{(vector) x2 of segment}

\item{y2}{(vector) y2 of segment}
}
\value{
(vector) distance to segment (dist),
                 derivate in x (dx),
                 derivate in y (dy)
}
\description{
Give distance a point to segment and derivate in x and y
}
