% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{extract_element_}
\alias{extract_element_}
\title{Extract elements (generic)}
\usage{
extract_element_(object)
}
\arguments{
\item{object}{(Spatial*DataFrame): Typing polygon or line landscape}
}
\value{
(matrix) M*5 (x1, y1, x2, y2, id_type)
}
\description{
Extract all element of typed polygon or line to create list 
             of matrix containing coordonnate of element with parameter of 
             potential function
}
