% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{extract_elements}
\alias{extract_elements}
\title{Extract elements}
\usage{
extract_elements(x, y, land_poly, land_lines)
}
\arguments{
\item{x}{(double) position in x}

\item{y}{(double) position in y}

\item{land_poly}{(SpatialPolygonDataFrame): Typing polygon landscape}

\item{land_lines}{(SpatialLinesDataFrame): Typing line landscape}
}
\value{
(matrix) M*4 (dist, dx, dy, id_type)
}
\description{
Extract all element of typed landscape to create list of matrix
             containing coordonnate of element with parameter of potential
             function
}
