% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{extract_lines}
\alias{extract_lines}
\title{Extract lines of spatialpolygons}
\usage{
extract_lines(landscape)
}
\arguments{
\item{landscape}{(SpatialPolygons*)}
}
\value{
SpatialLinesDataFrame
}
\description{
Extract lines of spatialpolygons
}
