% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{gen_land}
\alias{gen_land}
\title{Generate landscape}
\usage{
gen_land(nb_poly = 10, width = 60, height = 60)
}
\arguments{
\item{nb_poly}{: number of polygon}

\item{width}{: width of landscape}

\item{height}{: height of landscape}
}
\value{
SpatialPolygonDataFrame : landscape with convex ploygons
}
\description{
Generer un paysage de polygone convexe
}
