% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{get_neighbours}
\alias{get_neighbours}
\title{get neighbours}
\usage{
get_neighbours(landscape)
}
\arguments{
\item{landscape}{(SpatialPolygons*): the landscape}
}
\value{
list of list: one row by polygon. In each list of neighbours polygons
}
\description{
give neighbours of each polygon
}
