% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{getCoordsSpatial_}
\alias{getCoordsSpatial_}
\title{get coordonnate Spatial*}
\usage{
getCoordsSpatial_(object, name_elements, id)
}
\arguments{
\item{object}{(Spatial*)}

\item{name_elements}{(vector str): two name to attain coords}

\item{id}{(str): id of element}
}
\value{
(matrix X*2): coordinates of element
}
\description{
get coordonnate of element of Spatial*
}
