% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{getCoordsSpatialLines}
\alias{getCoordsSpatialLines}
\title{get coordonnate SpatialLines}
\usage{
getCoordsSpatialLines(my_lines, id)
}
\arguments{
\item{id}{(int): id of polygon}

\item{landscape}{(SpatialPolygons)}
}
\value{
(matrix X*2): coordinates of polygon
}
\description{
get coordonnate of line in Spatiallines
}
