% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{getCoordsSpatialPolygons}
\alias{getCoordsSpatialPolygons}
\title{get coordonnate SpatialPolygons}
\usage{
getCoordsSpatialPolygons(landscape, id)
}
\arguments{
\item{landscape}{(SpatialPolygons)}

\item{id}{(int): id of polygon}
}
\value{
(matrix X*2): coordinates of polygon
}
\description{
get coordonnate of polygon in SpatialPolygons
}
