% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{getIdsSpatial_}
\alias{getIdsSpatial_}
\title{Get identifiant of Spatial*}
\usage{
getIdsSpatial_(object, name_element)
}
\arguments{
\item{object}{(Spatial*): object Spatial*}

\item{name_element}{(str):}
}
\value{
list of id
}
\description{
Give Id of all elements
}
