% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{getIdsSpatialLines}
\alias{getIdsSpatialLines}
\title{Get identifiant of SpatialLines}
\usage{
getIdsSpatialLines(sp_line)
}
\arguments{
\item{sp_line}{(SpatialLines*): the landscape}
}
\value{
list of id
}
\description{
Give Id of all lines
}
