% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_functions.R
\name{getIdsSpatialPolygons}
\alias{getIdsSpatialPolygons}
\title{Get identifiant of SpatialPolygons}
\usage{
getIdsSpatialPolygons(land)
}
\arguments{
\item{land}{(SpatialPolygons*): the landscape}
}
\value{
list of id
}
\description{
Give Id of all polygons
}
