% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_potential_func}
\alias{grad_potential_func}
\title{gradient potential function}
\usage{
grad_potential_func(alpha_t, beta, dist, puiss, deriv)
}
\arguments{
\item{alpha_t}{(double)}

\item{beta}{(double) spatial display of potential}

\item{dist}{(double) distance}

\item{puiss}{(double)}

\item{deriv}{(double)}
}
\value{
(double) gradiant of potential
}
\description{
Fonction gradient du potentiel précédent
}
