% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{Interact}
\alias{Interact}
\title{Interact class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Interact
}
\description{
Define Interact of type(undefined) with a other
(defined : id_interact)
}
\keyword{datasets}
