% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{is_PotentialPolygon}
\alias{is_PotentialPolygon}
\title{Is PotentialPolygon}
\usage{
is_PotentialPolygon(x)
}
\arguments{
\item{x}{element}
}
\value{
boolean
}
\description{
Indicate if element is a PotentialPolygon
}
