% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{is_TypeInteractModel}
\alias{is_TypeInteractModel}
\title{Is TypeInteractModel}
\usage{
is_TypeInteractModel(x)
}
\arguments{
\item{x}{element}
}
\value{
boolean
}
\description{
Indicate if element is a TypeInteractModel
}
