% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{next_coord}
\alias{next_coord}
\title{next coord}
\usage{
next_coord(nv_coords_point, ui_width, ui_heigth, d_sigma, nm_coords_element,
  nm_info_type, time_step)
}
\arguments{
\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{d_sigma}{(double) repulsif effect adaptor}

\item{nm_coords_element}{(NumericMatrix) all coordinates of segment 
Matrix N*5 (x1,y1,x2,y2,Id_type)}

\item{nm_info_type}{(NumericMatrix) descript each type
Matrix M*5 (Id, rep (-1) or attr(1), 
            alpha, beta, power)}

\item{time_step}{(double) step of time}

\item{ui_land_width}{(unsigned int) width of landscape.}

\item{ui_land_heigth}{(unsigned int) heigth of landscape.}
}
\value{
(NumericVector) next coord in x and y (['x'];['y'])
}
\description{
calculate next coordinates
}
