% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{plot_potential}
\alias{plot_potential}
\title{plot potential}
\usage{
plot_potential(landscape, land_component, precision = 1)
}
\arguments{
\item{landscape}{(SpatialPolygon): landscape}

\item{land_component}{(list of 2 matrix) 
repulsive : M*7 (x1, y1, x2, y2, alpha, beta, power)
attractive : M*7 (x1, y1, x2, y2, alpha, beta, power)}

\item{precision}{(float) plot precision}
}
\description{
plot potential on landscape
}
