% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{potential_effect}
\alias{potential_effect}
\title{effect potential}
\usage{
potential_effect(nm_dist_element, nm_info_type)
}
\arguments{
\item{nm_dist_element}{(NumericMatrix) all distance and gradiant of segment 
and polygon 
(matrix) M*4 (dist, dx, dy, id_type)}

\item{nm_info_type}{(NumericMatrix) descript each type
Matrix M*5 (Id, rep (-1) or attr(1), 
            alpha, beta, power)}
}
\value{
(NumericVector) effect potential in x and y (['x'];['y'])
}
\description{
Calcualte gradient potential value for a x and y
}
