% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{potential_func}
\alias{potential_func}
\title{potential function}
\usage{
potential_func(alpha_t, beta, dist, power)
}
\arguments{
\item{alpha_t}{(double)}

\item{beta}{(double) spatial display of potential}

\item{dist}{(double) distance}

\item{power}{(double)}
}
\value{
(double) potential
}
\description{
potential function
}
