% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{potential_value}
\alias{potential_value}
\title{value potential}
\usage{
potential_value(nm_coords_element, nv_coords_point, bound, b_sum_sub)
}
\arguments{
\item{nm_coords_element}{(NumericMatrix) all coordinates of segment 
Matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
alpha_a (double) maximum potential 
amplitude,
beta (double) spatial display of 
potential,
and power (double)}

\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{bound}{(NumericVector) limit of value potential ("min", "max")}

\item{b_sum_sub}{(bool) use sum potential(true) or sub potential(false) 
(not effect on bound)}
}
\value{
(double) value potential
}
