% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\docType{data}
\name{PotentialPolygon}
\alias{PotentialPolygon}
\title{PotentialPolygon class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PotentialPolygon
}
\description{
Define the potential function of polygon
}
\section{methods}{

\itemize{
\item{PotentialPolygon$new(id, str_func)}{
\describe{initialize PotentialPolygon object
\itemize{
\item{id}{(str) identifiant of polygon}
\item{str_func}{(litteral potential function (eg:"5*exp(-x*2)")) }
\item{(int)}{0 : sucess,
             1 : impossible to transform str_func in function,
             2 : impossible to create derivate in x,
             3 : impossible to create derivate in y}
}}}
\item{$get_id()}{
\describe{Give id of polygone
\itemize{
\item{(characters)}{identifiant of polygon}
}}}
\item{$get_potential()}{
\describe{Give potential function
\itemize{
\item{(function)}{potential fucncion}
}}}
\item{$get_dx()}{
\describe{Give derivate of potential function in x
\itemize{
\item{(function)}{derivate of potential function in x}
}}}
\item{$get_dy()}{
\describe{Give derivate of potential function in y
\itemize{
\item{(function)}{derivate of potential function in y}
}}}
\item{$set_id(id)}{
\describe{Modify id of polygon
\itemize{
\item{id}{(characters) identifiant of polygon}
}}}
\item{$set_potential(str_func)}{
\describe{Modify potential function and derivate in x and y
\itemize{
\item{str_func}{(str) litteral potential function (eg:"5*exp(-x*2)")}
\item{(int)}{0 : sucess,
             1 : impossible to transform str_func in function,
             2 : impossible to create derivate in x,
             3 : impossible to create derivate in y}
}}}
}
}

\keyword{datasets}
