% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{potentiel_0}
\alias{potentiel_0}
\title{potentiel function define by user}
\usage{
potentiel_0(a, b, c, x0, y0, x1, y1)
}
\arguments{
\item{a}{(float): parameter of function}

\item{b}{(float): parameter of function}

\item{x0}{(float): abscisse of first point of line}

\item{y0}{(float): ordonate of first point of line}

\item{x1}{(float): abscisse of second point of line}

\item{y1}{(float): ordonate of second point of line}

\item{b}{(float): parameter of function}
}
\value{
(string): function expression
}
\description{
exponentiel decroissante
}
