% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repulsive_value}
\alias{repulsive_value}
\title{repulsive value}
\usage{
repulsive_value(nm_coords_rep, nv_coords_point)
}
\arguments{
\item{nm_coords_rep}{(NumericMatrix) all coordinates of repulsif segment 
Matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
   alpha_a (double) maximum potential 
   amplitude of repulsivity
   beta (double) spatial display of 
   potential, and power (double)}

\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}
}
\value{
(double) repulsif value of potential
}
\description{
Give value of potential on x and y of all replusive elements
}
