% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{TypeInteract}
\alias{TypeInteract}
\title{TypeInteract class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TypeInteract
}
\description{
Define interaction of several neightbour type on host type
}
\section{methods}{

\itemize{
\item{TypeInteract$new(id, name, func_agglo, interacts)}{
\describe{Create TypeInteract object
\itemize{
\item{id}{(integer) id of host type}
\item{name}{(character) name of type }
\item{func_agglo}{(function) function define how aggregate interactions}
\item{interacts}{(list of Interact) list of interaction of type with
                 other}
}}}
\item{$get_id()}{
\describe{Give identifiant of host type 
\itemize{
\item{(integer)}{ id of host type }
}}}
\item{$get_name()}{
\describe{Give name of host type
\itemize{
\item{(character)}{ name of host type }
}}}
\item{$get_func_agglo()}{
\describe{Give function of aggregate
\itemize{
\item{(function)}{ function define how aggregate interactions}
}}}
\item{$get_interact(id)}{
\describe{Give a interaction of neightbour type on host type
\itemize{
\item{id}{ (integer) id neightbour type}
\item{(Interact)}{ interaction}
}}}
\item{$get_id_interacts()}{
\describe{Give the list of id of neightbour type in interaction
\itemize{
\item{(list of integer)}{ list of id neightbour type}
}}}
\item{$get_function_interact(id)}{
\describe{Give interaction function 
\itemize{
\item{id}{ (integer) id neightbour type}
\item{(function)}{ interaction function}
}}}
\item{$get_params(id)}{
\describe{Give parameters of interaction function
\itemize{
\item{id}{ (integer) id neightbour type}
\item{(list)}{ list of constant parameters of function}
}}}
\item{$set_id(id)}{
\describe{Define identifiant of type
\itemize{
\item{id}{(integer) id of host type }
}}}
\item{$set_name(name)}{
\describe{Define name of host type
\itemize{
\item{name}{(character) name of host type }
}}}
\item{$set_func_agglo(func_agglo)}{
\describe{Define function of aggregate
\itemize{
\item{func_agglo}{(function) function define how aggregate interactions}
}}}
\item{$set_interact(interact)}{
\describe{Add or modify a interaction of neightbour type on host type
\itemize{
\item{interact}{(Interact) interaction of neightbour type}
\item{(integer)}{ 0: success, 1: fail}
}}}
\item{$set_func_interact(id, func)}{
\describe{Modify interaction function of one interaction
\itemize{
\item{id}{ (integer) id neightbour type}
\item{func}{ (function) interaction function}
\item{(integer)}{ 0: success, 1: fail}
}}}
\item{$set_param(id, param)}{
\describe{Modify parameters of interaction function of one interaction
\itemize{
\item{id}{ (integer) id neightbour type}
\item{param}{ (list) list of constant parameters of function}
\item{(integer)}{ 0: success, 1: fail}
}}}
\item{$remove_interact(id)}{
\describe{remove a interaction of neightbour typer
\itemize{
\item{id}{ (integer) id neightbour type}
\item{(integer)}{ 0: success, 1: fail}
}}}
}
}

\keyword{datasets}
