% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{all_effect}
\alias{all_effect}
\title{all effect}
\usage{
all_effect(nv_coords_point, ui_width, ui_heigth, d_sigma, nm_coords_attrac,
  alpha_a, nm_coords_rep, alpha_r, time_step)
}
\arguments{
\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{d_sigma}{(double) repulsif effect adaptor}

\item{nm_coords_attrac}{(NumericMatrix) all coordinates of attractive 
segment matrix N*4 (x1,y1,x2,y2)}

\item{alpha_a}{(double) maximum potential amplitude of attrativity}

\item{nm_coords_rep}{(NumericMatrix) all coordinates of repulsif segment 
Matrix N*4 (x1,y1,x2,y2)}

\item{alpha_r}{(double) maximum potential amplitude of repulsivity}

\item{time_step}{(double) step of time}

\item{ui_land_width}{(unsigned int) width of landscape.}

\item{ui_land_heigth}{(unsigned int) heigth of landscape.}
}
\value{
(NumericVector) effect in x and y (['x'];['y'])
}
\description{
Give complete effect of landscape on x and y
}
