% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{extract_element_}
\alias{extract_element_}
\title{Extract elements (generic)}
\usage{
extract_element_(x, y, object)
}
\arguments{
\item{x}{(double) position in x}

\item{y}{(double) position in y}

\item{object}{(Spatial*DataFrame): Typing polygon or line landscape}
}
\value{
(matrix) M*4 (dist, dx, dy, id_type)
}
\description{
Extract all element of typed polygon or line to create list 
             of matrix containing coordonnate of element with parameter of 
             potential function
}
