% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{extract_element_}
\alias{extract_element_}
\title{Extract elements (generic)}
\usage{
extract_element_(object, info_types)
}
\arguments{
\item{object}{(Spatial*DataFrame): Typing polygon or line landscape}

\item{info_type}{(matrix): N*5 (id, repulsive=0 or attractive=1, alpha, beta,
power) where N number of type.}
}
\value{
(matrix) M*7 (x1, y1, x2, y2, alpha, beta,
                      power)
}
\description{
Extract all element of typed polygon or line to create list 
             of matrix containing coordonnate of element with parameter of 
             potential function
}
