% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{extract_elements}
\alias{extract_elements}
\title{Extract elements}
\usage{
extract_elements(land_poly, land_lines, info_types)
}
\arguments{
\item{land_poly}{(SpatialPolygonDataFrame): Typing polygon landscape}

\item{land_lines}{(SpatialLinesDataFrame): Typing line landscape}

\item{info_type}{(matrix): N*5 (id, repulsive=0 or attractive=1, alpha, beta,
power) where N number of type.}
}
\value{
(list of 2 matrix) 
           repulsive : M*7 (x1, y1, x2, y2, alpha, beta, power)
           attractive : M*7 (x1, y1, x2, y2, alpha, beta, power)
}
\description{
Extract all element of typed landscape to create list of matrix
             containing coordonnate of element with parameter of potential
             function
}
