% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{extract_elements}
\alias{extract_elements}
\title{Extract elements}
\usage{
extract_elements(land_poly, land_lines)
}
\arguments{
\item{land_poly}{(SpatialPolygonDataFrame): Typing polygon landscape}

\item{land_lines}{(SpatialLinesDataFrame): Typing line landscape}
}
\value{
(matrix) M*5 (x1, y1, x2, y2, id_type)
}
\description{
Extract all element of typed landscape to create list of matrix
             containing coordonnate of element with parameter of potential
             function
}
