% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{next_coord}
\alias{next_coord}
\title{next coord}
\usage{
next_coord(nv_coords_point, ui_width, ui_heigth, d_sigma, nm_coords_attrac,
  nm_coords_rep, time_step)
}
\arguments{
\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{d_sigma}{(double) repulsif effect adaptor}

\item{nm_coords_attrac}{(NumericMatrix) all coordinates of attractive 
segment matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
alpha_a (double) maximum potential 
amplitude of repulsivity
beta (double) spatial display of 
potential, and power (double)}

\item{nm_coords_rep}{(NumericMatrix) all coordinates of repulsif segment 
Matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
   alpha_a (double) maximum potential 
   amplitude of attractivity
   beta (double) spatial display of 
   potential, and power (double)}

\item{time_step}{(double) step of time}

\item{ui_land_width}{(unsigned int) width of landscape.}

\item{ui_land_heigth}{(unsigned int) heigth of landscape.}
}
\value{
(NumericVector) next coord in x and y (['x'];['y'])
}
\description{
calculate next coordinates
}
