% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{next_coord}
\alias{next_coord}
\title{next coord}
\usage{
next_coord(x, y, land_poly, land_lines, info_type, sigma, time_step)
}
\arguments{
\item{x}{(double) position in x}

\item{y}{(double) position in y}

\item{land_poly}{(SpatialPolygonDataFrame): Typing polygon landscape}

\item{land_lines}{(SpatialLinesDataFrame): Typing line landscape}

\item{info_type}{(matrix): N*5 (id, repulsive=0 or attractive=1, alpha, beta,
power) where N number of type.}

\item{sigma}{(double) repulsif effect adaptor}

\item{time_step}{(double) step of time}
}
\value{
(NumericVector) next coord in x and y (['x'];['y'])
}
\description{
calculate next coordinates
}
