% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{plot_potential}
\alias{plot_potential}
\title{plot potential}
\usage{
plot_potential(land_poly, land_lines, info_type, precision = 1)
}
\arguments{
\item{land_poly}{(SpatialPolygonDataFrame): Typing polygon landscape}

\item{land_lines}{(SpatialLinesDataFrame): Typing line landscape}

\item{info_type}{(matrix): N*5 (id, repulsive=0 or attractive=1, alpha, beta,
power) where N number of type.}

\item{precision}{(float) plot precision}
}
\value{
(matrix) matrix of potential
}
\description{
plot potential on landscape
}
