% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{potential_effect}
\alias{potential_effect}
\title{effect potential}
\usage{
potential_effect(nm_coords_element, nv_coords_point, bound, alpha, beta, power,
  b_sum_sub)
}
\arguments{
\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{bound}{(NumericVector) limit of effect potential ("min", "max")}

\item{alpha}{(double) potential amplitude}

\item{beta}{(double) spatial display of potential}

\item{power}{(double)}

\item{b_sum_sub}{(bool) use +gradiant(true) or -gradiant(false) (not effect 
on bound)}

\item{nm_coords_rep}{(NumericMatrix) all coordinates of repulsif segment 
Matrix N*4 (x1,y1,x2,y2)}
}
\value{
(NumericVector) effect potential in x and y (['x'];['y'])
}
