% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{potential_effect}
\alias{potential_effect}
\title{effect potential}
\usage{
potential_effect(nm_coords_element, nv_coords_point, bound, b_sum_sub)
}
\arguments{
\item{nm_coords_element}{(NumericMatrix) all coordinates of segment 
Matrix N*7 (x1,y1,x2,y2,
alpha, beta, power) where
alpha_a (double) maximum potential 
amplitude,
beta (double) spatial display of 
potential,
and power (double)}

\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{bound}{(NumericVector) limit of effect potential ("min", "max")}

\item{b_sum_sub}{(bool) use +gradiant(true) or -gradiant(false) (not effect 
on bound)}
}
\value{
(NumericVector) effect potential in x and y (['x'];['y'])
}
