% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{potential_value}
\alias{potential_value}
\title{value potential}
\usage{
potential_value(nv_coords_point, nm_coords_element, nm_info_type)
}
\arguments{
\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{nm_coords_element}{(NumericMatrix) all coordinates of segment 
Matrix N*5 (x1,y1,x2,y2,Id_type)}

\item{nm_info_type}{(NumericMatrix) descript each type
Matrix M*5 (Id, rep (-1) or attr(1), 
            alpha, beta, power)}
}
\value{
(double) value potential
}
