% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{potential_value}
\alias{potential_value}
\title{value potential}
\usage{
potential_value(nm_dist_element, nm_info_type)
}
\arguments{
\item{nm_dist_element}{(NumericMatrix) all distance and gradiant of segment 
and polygon 
(matrix) M*4 (dist, dx, dy, id_type)}

\item{nm_info_type}{(NumericMatrix) descript each type
Matrix M*5 (Id, rep (-1) or attr(1), 
            alpha, beta, power)}
}
\value{
(double) value potential
}
\description{
Calcualte potential value for a x and y
}
