% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\docType{data}
\name{PotentialLandscape}
\alias{PotentialLandscape}
\title{PotentialLandscape}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PotentialLandscape
}
\description{
Define a landscape with potential function of each polygon
}
\keyword{datasets}
