% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repulsive_effect}
\alias{repulsive_effect}
\title{repulsive effect}
\usage{
repulsive_effect(nm_coords_rep, nv_coords_point, alpha1)
}
\arguments{
\item{nm_coords_rep}{(NumericMatrix) all coordinates of repulsif segment 
Matrix N*4 (x1,y1,x2,y2)}

\item{nv_coords_point}{(NumericVector) coordinates of point (x,y)}

\item{alpha1}{(double) maximum potential amplitude}
}
\value{
(NumericVector) repulsif effect in x and y (['x'];['y'])
}
\description{
Give effect on x and y of all replusive elements
}
