% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{TypeInteract}
\alias{TypeInteract}
\title{TypeInteract class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TypeInteract
}
\description{
Define Interact of type with others
}
\keyword{datasets}
